feat(DeeplinkService): If app started with a public share link before authentication it is replayed after
