+++ kDriveCore/Utils/Deeplinks/DeeplinkService.swift
@@ -17,45 +17,26 @@
  */
 
 import Foundation
-import SwiftRegex
-
-public struct PublicShareLink {
-    public static let parsingRegex = Regex(pattern: #"^/app/share/([0-9]+)/([a-z0-9-]+)$"#)
-
-    public let publicShareURL: URL
-    public let shareLinkUid: String
-    public let driveId: Int
-
-    public init?(publicShareURL: URL) async {
-        guard let components = URLComponents(url: publicShareURL, resolvingAgainstBaseURL: true) else {
-            return nil
-        }
-
-        let path = components.path
-        guard let matches = Self.parsingRegex?.matches(in: path) else {
-            return nil
-        }
-
-        guard let firstMatch = matches.first,
-              let driveId = firstMatch[safe: 1],
-              let driveIdInt = Int(driveId),
-              let shareLinkUid = firstMatch[safe: 2] else {
-            return nil
-        }
-
-        self.driveId = driveIdInt
-        self.shareLinkUid = shareLinkUid
-        self.publicShareURL = publicShareURL
-    }
-}
 
 public protocol DeeplinkServiceable: AnyObject {
     func setLastPublicShare(_ link: PublicShareLink)
+    func processDeeplinksPostAuthentication()
 }
 
 public class DeeplinkService: DeeplinkServiceable {
     var lastPublicShareLink: PublicShareLink?
     public func setLastPublicShare(_ link: PublicShareLink) {
         lastPublicShareLink = link
+    }
+
+    public func processDeeplinksPostAuthentication() {
+        guard let lastPublicShareLink else {
+            return
+        }
+
+        Task {
+            await UniversalLinksHelper.processPublicShareLink(lastPublicShareLink)
+            self.lastPublicShareLink = nil
+        }
     }
 }
