+++ kDriveCore/Utils/Deeplinks/PublicShareLink.swift
@@ -1,4 +1,3 @@
-//
 /*
  Infomaniak kDrive - iOS App
  Copyright (C) 2024 Infomaniak Network SA
@@ -18,3 +17,34 @@
  */
 
 import Foundation
+import SwiftRegex
+
+public struct PublicShareLink: Sendable {
+    public static let parsingRegex = Regex(pattern: #"^/app/share/([0-9]+)/([a-z0-9-]+)$"#)
+
+    public let publicShareURL: URL
+    public let shareLinkUid: String
+    public let driveId: Int
+
+    public init?(publicShareURL: URL) async {
+        guard let components = URLComponents(url: publicShareURL, resolvingAgainstBaseURL: true) else {
+            return nil
+        }
+
+        let path = components.path
+        guard let matches = Self.parsingRegex?.matches(in: path) else {
+            return nil
+        }
+
+        guard let firstMatch = matches.first,
+              let driveId = firstMatch[safe: 1],
+              let driveIdInt = Int(driveId),
+              let shareLinkUid = firstMatch[safe: 2] else {
+            return nil
+        }
+
+        self.driveId = driveIdInt
+        self.shareLinkUid = shareLinkUid
+        self.publicShareURL = publicShareURL
+    }
+}
